requirements.txt
fastmcp>=0.3.0
anyio>=4.0.0
httpx>=0.27.0
mcp
openai
setuptools
python-dotenv